app/requirements.txt
Django==3.2
djangorestframework==3.13.1
simplejson==3.17.6
requests==2.27.1
psycopg2-binary==2.8.6
